Split test project into unit, integration and functional projects
