+++ OfficeRibbonXEditor.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.29318.209
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OfficeRibbonXEditor", "src\OfficeRibbonXEditor\OfficeRibbonXEditor.csproj", "{87EFED6B-45A1-425F-B63B-5894E4FC0C03}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OfficeRibbonXEditorTests", "tests\OfficeRibbonXEditorTests\OfficeRibbonXEditorTests.csproj", "{D4F06196-AE8D-4DF5-BE6B-49490919A67B}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{62BC0A54-36D3-47DD-8A1B-44AE03DDB305}"
 	ProjectSection(SolutionItems) = preProject
 		azure-pipelines.yml = azure-pipelines.yml
@@ -33,6 +31,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "build", "build", "{D8B42177
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{5FB6177B-63FD-4817-A7ED-D1FA2D6EBEAF}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UnitTests", "tests\UnitTests\UnitTests.csproj", "{92E4D5A6-E84D-4EBD-867B-C5A2994855E5}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntegrationTests", "tests\IntegrationTests\IntegrationTests.csproj", "{0F75E691-6B9B-4092-B551-05D49EF445F4}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FunctionalTests", "tests\FunctionalTests\FunctionalTests.csproj", "{ACF85006-A9D7-4B30-97C8-C964EA6494E4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -49,14 +53,6 @@ Global
 		{87EFED6B-45A1-425F-B63B-5894E4FC0C03}.Release|Any CPU.Build.0 = Release|Any CPU
 		{87EFED6B-45A1-425F-B63B-5894E4FC0C03}.Release|x86.ActiveCfg = Release|Any CPU
 		{87EFED6B-45A1-425F-B63B-5894E4FC0C03}.Release|x86.Build.0 = Release|Any CPU
-		{D4F06196-AE8D-4DF5-BE6B-49490919A67B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D4F06196-AE8D-4DF5-BE6B-49490919A67B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D4F06196-AE8D-4DF5-BE6B-49490919A67B}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D4F06196-AE8D-4DF5-BE6B-49490919A67B}.Debug|x86.Build.0 = Debug|Any CPU
-		{D4F06196-AE8D-4DF5-BE6B-49490919A67B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D4F06196-AE8D-4DF5-BE6B-49490919A67B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D4F06196-AE8D-4DF5-BE6B-49490919A67B}.Release|x86.ActiveCfg = Release|Any CPU
-		{D4F06196-AE8D-4DF5-BE6B-49490919A67B}.Release|x86.Build.0 = Release|Any CPU
 		{414E2267-764A-4E61-AA5E-F25F407DFAFD}.Debug|Any CPU.ActiveCfg = Debug|x86
 		{414E2267-764A-4E61-AA5E-F25F407DFAFD}.Debug|x86.ActiveCfg = Debug|x86
 		{414E2267-764A-4E61-AA5E-F25F407DFAFD}.Debug|x86.Build.0 = Debug|x86
@@ -80,16 +76,42 @@ Global
 		{BD57B012-822B-4C74-95B8-8254B320CAF3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BD57B012-822B-4C74-95B8-8254B320CAF3}.Release|x86.ActiveCfg = Release|Any CPU
 		{BD57B012-822B-4C74-95B8-8254B320CAF3}.Release|x86.Build.0 = Release|Any CPU
+		{92E4D5A6-E84D-4EBD-867B-C5A2994855E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{92E4D5A6-E84D-4EBD-867B-C5A2994855E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{92E4D5A6-E84D-4EBD-867B-C5A2994855E5}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{92E4D5A6-E84D-4EBD-867B-C5A2994855E5}.Debug|x86.Build.0 = Debug|Any CPU
+		{92E4D5A6-E84D-4EBD-867B-C5A2994855E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{92E4D5A6-E84D-4EBD-867B-C5A2994855E5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{92E4D5A6-E84D-4EBD-867B-C5A2994855E5}.Release|x86.ActiveCfg = Release|Any CPU
+		{92E4D5A6-E84D-4EBD-867B-C5A2994855E5}.Release|x86.Build.0 = Release|Any CPU
+		{0F75E691-6B9B-4092-B551-05D49EF445F4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0F75E691-6B9B-4092-B551-05D49EF445F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0F75E691-6B9B-4092-B551-05D49EF445F4}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0F75E691-6B9B-4092-B551-05D49EF445F4}.Debug|x86.Build.0 = Debug|Any CPU
+		{0F75E691-6B9B-4092-B551-05D49EF445F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0F75E691-6B9B-4092-B551-05D49EF445F4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0F75E691-6B9B-4092-B551-05D49EF445F4}.Release|x86.ActiveCfg = Release|Any CPU
+		{0F75E691-6B9B-4092-B551-05D49EF445F4}.Release|x86.Build.0 = Release|Any CPU
+		{ACF85006-A9D7-4B30-97C8-C964EA6494E4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ACF85006-A9D7-4B30-97C8-C964EA6494E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ACF85006-A9D7-4B30-97C8-C964EA6494E4}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{ACF85006-A9D7-4B30-97C8-C964EA6494E4}.Debug|x86.Build.0 = Debug|Any CPU
+		{ACF85006-A9D7-4B30-97C8-C964EA6494E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ACF85006-A9D7-4B30-97C8-C964EA6494E4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ACF85006-A9D7-4B30-97C8-C964EA6494E4}.Release|x86.ActiveCfg = Release|Any CPU
+		{ACF85006-A9D7-4B30-97C8-C964EA6494E4}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{87EFED6B-45A1-425F-B63B-5894E4FC0C03} = {017FBD33-4473-4DD2-ADEA-CAA00B695055}
-		{D4F06196-AE8D-4DF5-BE6B-49490919A67B} = {5FB6177B-63FD-4817-A7ED-D1FA2D6EBEAF}
 		{414E2267-764A-4E61-AA5E-F25F407DFAFD} = {017FBD33-4473-4DD2-ADEA-CAA00B695055}
 		{22AE2386-60F1-476E-9303-61CDB0AAC4CF} = {017FBD33-4473-4DD2-ADEA-CAA00B695055}
 		{BD57B012-822B-4C74-95B8-8254B320CAF3} = {017FBD33-4473-4DD2-ADEA-CAA00B695055}
+		{92E4D5A6-E84D-4EBD-867B-C5A2994855E5} = {5FB6177B-63FD-4817-A7ED-D1FA2D6EBEAF}
+		{0F75E691-6B9B-4092-B551-05D49EF445F4} = {5FB6177B-63FD-4817-A7ED-D1FA2D6EBEAF}
+		{ACF85006-A9D7-4B30-97C8-C964EA6494E4} = {5FB6177B-63FD-4817-A7ED-D1FA2D6EBEAF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8E5E4631-F9D6-43A1-B677-FD1B0C599606}
