+++ tests/FunctionalTests/Resources/undo.png
