--- tests/OfficeRibbonXEditorTests/Models/OfficePartTests.cs
@@ -1,10 +1,0 @@
-﻿using NUnit.Framework;
-
-namespace OfficeRibbonXEditor.Models
-{
-    [TestFixture]
-    public class OfficePartTests
-    {
-
-    }
-}
