--- tests/OfficeRibbonXEditorTests/Properties/AssemblyInfo.cs
@@ -1,16 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("OfficeRibbonXEditorTests")]
-[assembly: AssemblyDescription("Tests for the OfficeRibbonXEditor project")]
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("d4f06196-ae8d-4df5-be6b-49490919a67b")]
